requirements.txt
langchain_core==0.3.51
langchain_openai==0.3.12
langgraph==0.3.25
python-dotenv==1.1.0
Requests==2.32.3
setuptools==78.1.0
